requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
crispy-bootstrap5==0.7
deprecation==2.1.0
Django==4.1.6
django-bootstrap-v5==1.0.11
django-crispy-forms==1.14.0
django-decouple==2.1
django-recaptcha==3.0.0
django-tinymce==3.5.0
packaging==23.0
Pillow==9.4.0
psycopg2==2.9.5
PyJWT==2.6.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.7
